Iterating/building on ideas for asset management sqlite integration asset import
got to the point where I am trying to make an extension to
sqlite to support the hashing scheme used in the code
which would allow sql queries using the same hashes.

the rot13 example works, but I wanted to make it persistent.

dll stuff was a pain. Got caught out with a cpp extension instead of c
which meant the dlls entry point could not be found.

Looks as though sqlite is plenty performant for an engine. Why don't
all engines use it?

Will make an abstraction layer so you can pick your db implementation.